requirements.txt
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
greenlet==3.0.3
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.0
pluggy==1.5.0
PyMySQL==1.1.1
python-dotenv==1.0.1
SQLAlchemy==2.0.30
typing_extensions==4.12.0
urllib3==2.2.1
Werkzeug==3.0.3
